+++ NEWS.md
@@ -1,3 +1,8 @@
+# normalr 0.0.2
+
+* add the shiny app using `normalrShiny()`.
+* minor fix on the negative data with negative lambda.
+
 # normalr 0.0.1
 
 * First release with three main functions: 
@@ -5,7 +10,5 @@
   * `normalise()` normalised a numeric vector with the specific lambda value.
   * `normaliseData()` applies the normalisation on a dataset with multiple continous variables.
 
-# normalr 0.0.2
 
-* add the shiny app using `normalrShiny()`.